social-crawler/pyproject.toml
[project]
name = "social-crawler-demo"
version = "0.1.0"
description = "Public capability demo: pure asyncio + Playwright crawler for Facebook / Twitter / Instagram / TikTok with CDP attach and click-flow navigation. Subdirectory of freelance-demos."
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "playwright>=1.45",
    "python-dotenv>=1.0",
    "click>=8.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/social_crawler"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
